refactor: #343 detail component folder
