+++ resources/js/Components/Detail/CharacterDetailBox.test.js
