+++ resources/js/Components/Detail/CharacterDetailBox.tsx
