+++ resources/js/Components/Detail/ItemDetailBox.test.js
