+++ resources/js/Components/Detail/ItemDetailBox.tsx
