+++ resources/js/Components/Detail/index.ts
@@ -1,1 +1,2 @@
 export {default as CharacterDetailBox} from './CharacterDetailBox'
+export {default as ItemDetailBox} from './ItemDetailBox'
