--- resources/js/Components/ItemDetail/index.ts
@@ -1,1 +1,0 @@
-export {default as ItemDetailBox} from './ItemDetailBox'
